// frontend-old/lib/types.ts
export type CartItem = {
  id: string,
  name: string,
  price: number,
  quantity: number,
  product: Product

}

export type Product = {
  id: string,
  price: number,
  name: string,
  slug: string,
  status: string,
  stockCount: string,
  excerpt:string
  description: {
    document:any
  },
  photo: Photo,
  image:string,
  stripeProductId: string,
  stripePriceId: string,
  tags: Tag[],
  categories: Category[],
  dateCreated: string,
  dateModified: string,
  author:User,
}

export type ProductImage = {
  image: any,
  url: string,
  altText: string,
  filename: string,
  product: Product,
  subscription: any,
}

export type Photo = {
  id: string,
  altText: string,
  image: {
    url: string
  }
}

export type Orders = {
  Orders: [Order]
}

export type Order = {
  id: string,
  charge: string,
  total: number,
  createdAt: string,
  user: User,
  items: [OrderItem],
  itemsCount:number,
}

export type OrderItem = {
  id: string,
  name: string,
  description: string,
  price: number,
  quantity: number,
  productId: string,
  image:string,
  photo: Photo,
  dateCreated: string,
  dateModified: string,
}

export type User = {
  id: string,
  name?: string,
  nameLast?:string,
  image?:string,
  email?: string,
  phone?: string,
  password?: string,
  isAdmin?: boolean,
  isActive?: boolean,
  stripeCustomerId?: string,
  posts?: any[],
  pages?: any[],
  servicesProvided?: Service[],
  bookings?: any[],
  gigs?: any[],
  availability?: Availability[],
  cart?: CartItem[],
  createdAt?: string,
  products?: Product[],
  subscriptionPlans?: any[],
  subscriptions?: any[],
  orders?: OrderItem[],
  role?: {
    name:string,
    canManageCart?:boolean,
    canManageOrders?:boolean,
    canManageProducts?:boolean,
    canManageRoles?:boolean,
    canManageUsers?:boolean,
    canSeeOtherUsers?:boolean,
    canManageTickets?:boolean,
    canManageEvents?:boolean,
  }
  dateCreated?: string,
  dateModified?: string,
  tickets?:Ticket[],
}

export type Availability = {
  id: string,
  start: string,
  end: string,
  durationInHours: string,
  employee: User,
  type: string,
  status: string,
  dateCreated: string,
  dateModified: string,
}

export type Booking = {
  id: string,
  start: string,
  end: string,
  summary: string,
  durationInHours: string,
  service: Service,
  price: number,
  employees: User[],
  customer: User,
  email:string,
  phone:string,
  name:string,
  notes: string,
  dateCreated: string,
  dateModified: string,
  addons:Addon[],
  location:Location,
}

export type Category = {
  id: string,
  name: string,
  description: string,
  pages: Page[],
  posts: Post[],
  products: Product[],
  subscriptions: any[],
  services: Service[],
}

export type Tag = {
  id: string,
  name: string,
  pages: Page[],
  posts: Post[],
  products: Product[],
  subscriptions: any[],
  services: Service[],
}

export type Page = {
  id: string,
  title: string,
  slug: string,
  dateCreated: string,
  dateModified: string,
  status: string,
  template: string,
  pinned: number,
  excerpt: string,
  featured_image: string,
  featured_video: string,
  content: any,
  author: User,
  categories: Category[],
  tags: Tag[],
  allow_comments:boolean,
}

export type Post = {
  id: string,
  title: string,
  slug: string,
  dateCreated: string,
  dateModified: string,
  status: string,
  template: string,
  pinned: number,
  excerpt: string,
  featured_image: string,
  featured_video: string,
  content: any,
  author: User,
  categories: Category[],
  tags: Tag[],
  allow_comments:boolean,
}

export type Role = {
  id: string,
  name: string,
  assignedTo: User,
}

export type Service = {
  id: string,
  name: string,
  description: string,
  price: number,
  durationInHours: string,
  buisnessHourOpen: string,
  buisnessHourClosed: string,
  buisnessDays: number[],
  employees: User[],
  bookings: Booking[],
  status: string,
  categories: Category[],
  tags: Tag[],
  locations: Location[],
  addons: Addon[],
}
// ...
export type Addon = {
  id: string,
  name: string,
  description: string,
  price: number,
  services: Service[],
  bookings: Booking[],
  categories: Category[],
  tags: Tag[],
}

export type IDObj = {
  id: string,
}

export type Location = {
  id:string,
  name?: string,
  address?:string,
  rooms?:number,
  services?: Service[],
  bookings?:Booking[],
  events?:Event[],
  tags?: Tag[],
  categories?: Category[],
}

export type Event = {
  id?:string,
  summary?:string,
  location?:Location,
  start?: string,
  end?:string,
  price?:number,
  hosts?:User[],
  tickets?: Ticket[],
  seats?:number,
  image:string,
  photo?:string,
  description: {
    document: any,
  }
  excerpt:string,
  status?:string,
  dateCreated?:string,
  dateModified?:string,
  tags?: Tag[],
  categories?: Category[],
}

export type Ticket = {
  id:string,
  qrcode:string,
  event: Event,
  holder: User,
  status: string,
}